Fix C90-pedantic build errors: ssize_t and %zu
todo_api.h used ssize_t without including <sys/types.h>, so any
translation unit including it before <unistd.h> (e.g. urlencoded.c)
failed to compile. gcc 4.8.4's default C90 mode also rejects the C99
%zu length modifier under -pedantic; use %lu with an explicit cast
instead.

+++ sockets/response.c
@@ -46,7 +46,7 @@ void send_response(int fd, int code, const char *reason,
 	length_line[0] = '\0';
 	if (content_type)
 		snprintf(length_line, sizeof(length_line),
-			"Content-Length: %zu\r\n", body_len);
+			"Content-Length: %lu\r\n", (unsigned long)body_len);
 
 	header_len = snprintf(header, sizeof(header),
 		"HTTP/1.1 %d %s\r\n"
